Manually fixed glyphs in patched Source Code Pro
The glyphs that were added by fontpatcher.py were not positioned
correctly.  This was most noticable for the triangular separators. To
correct the glyphs I used the following modifications to fontpatcher.py:
- y_ratio *= 0.96
- y_diff += 300
The values are based on trial and error and what "looked" right, so
there is probably still room for improvement.