Add BDF and PCF files for terminus
Latter is really generated from the former, but first is a regular text file
where changes can be viewed (if you understand encoding; TXT is better in this
case) and second is what is installed.

Note: you can use my bitmap fontpatcher to get exactly the same fonts as I have
saved glyphs in glyphs/{size}/{codepoint}: this directory was created as
I assumed that version of the glyph adjusted by me is suitable for any font with
the same font size.