Use /bin/sh as the interpreter for installation script.
Users of *BSD systems will usually have bash installed under /usr/local or something of the sort.
There were two viable options here, the other is to use the `env` tool to determine the location
of bash. However, after linting this file with the sh interpreter installed on my system,
I discovered that this will do.